virtcca_deploy/requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
libvirt-python==11.10.0
MarkupSafe==3.0.3
packaging==25.0
psutil==7.2.1
requests==2.32.5
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
zope.event==6.1
zope.interface==8.2
